CSN: CSN-leap-platforms-572
Support for 64-bit UMF chunks in simulation.

+++ modules/bluespec/common/fpgaenv/physical-platform/physical-devices/unix-pipe/unix-pipe-device-bdpi.c
@@ -301,7 +301,7 @@ void pipe_read(unsigned int* resultptr, unsigned char handle)
         /* the following code is endian-agnostic */
         for (i = 0; i < BDPI_CHUNK_BYTES; i++)
         {
-            unsigned int byte = (unsigned int)channel->inputBuffer[channel->ibHead];
+            unsigned long long byte = (unsigned long long)channel->inputBuffer[channel->ibHead];
             retval |= (byte << (i * 8));
             channel->ibHead = channel->ibHead + 1;
         }
